feat: add dto file
