+++ sales-api/internal/events/usecase/list_events.go
@@ -2,23 +2,6 @@ package usecase
 
 import "github.com/dyhalmeida/ticket-full-cycle/sales-api/internal/events/domain"
 
-type EventDTO struct {
-	ID           string  `json:"id"`
-	Name         string  `json:"name"`
-	Location     string  `json:"location"`
-	Organization string  `json:"organization"`
-	Rating       string  `json:"rating"`
-	Date         string  `json:"date"`
-	Capacity     int     `json:"capacity"`
-	Price        float64 `json:"price"`
-	PartnerID    int     `json:"partner_id"`
-	ImageURL     string  `json:"image_url"`
-}
-
-type ListEventsOutputDTO struct {
-	Events []EventDTO `json:"events"`
-}
-
 type ListEventsUseCase struct {
 	eventRepository domain.EventRepositoryInterface
 }
